Feat: moved diffirent functionallity to diffierent files
